test/requirements.txt
async-timeout==4.0.2
Deprecated==1.2.13
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pymongo==4.2.0
pyparsing==3.0.9
redis==4.3.4
wrapt==1.14.1
